+++ src/main/java/step3/application/domain/RacingCars.java
@@ -1,7 +1,7 @@
 package step3.application.domain;
 
-import step3.application.domain.model.MovementLog;
-import step3.application.domain.model.OneMovementLog;
+import step3.application.domain.model.dto.MovementLog;
+import step3.application.domain.model.dto.OneMovementLog;
 import step3.application.generator.MovableGenerator;
 
 import java.util.List;
