+++ src/main/java/step3/application/domain/model/vo/Position.java
@@ -1,4 +1,4 @@
-package step3.application.domain.model;
+package step3.application.domain.model.vo;
 
 import error.ErrorMessage;
