+++ src/test/java/step3/application/domain/CarTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import step3.application.domain.model.MovementLog;
+import step3.application.domain.model.dto.MovementLog;
 
 import java.util.stream.Stream;
