+++ src/test/java/step3/application/domain/model/DriverTest.java
@@ -3,6 +3,8 @@ package step3.application.domain.model;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import step3.application.domain.model.dto.MovementLog;
+import step3.application.domain.model.vo.Driver;
 
 class DriverTest {
